+++ packages/circular-progress-bar/src/component.screenshots.test.tsx
@@ -20,7 +20,7 @@ describe(
                         title: ['', 'Title'],
                         subtitle: ['', 'SubTitle'],
                         view: ['positive', 'negative'],
-                        size: ['m', 'l'],
+                        size: ['s', 'm', 'l'],
                     },
                     size: { width: 200, height: 200 },
                 }),