+++ packages/intl-phone-input/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [6.0.4](https://github.com/alfa-laboratory/core-components/compare/@alfalab/core-components-intl-phone-input@6.0.3...@alfalab/core-components-intl-phone-input@6.0.4) (2021-12-01)
+
+**Note:** Version bump only for package @alfalab/core-components-intl-phone-input
+
+
+
+
+
+## [6.0.3](https://github.com/alfa-laboratory/core-components/compare/@alfalab/core-components-intl-phone-input@6.0.2...@alfalab/core-components-intl-phone-input@6.0.3) (2021-11-26)
+
+**Note:** Version bump only for package @alfalab/core-components-intl-phone-input
+
+
+
+
+
+## [6.0.2](https://github.com/alfa-laboratory/core-components/compare/@alfalab/core-components-intl-phone-input@6.0.1...@alfalab/core-components-intl-phone-input@6.0.2) (2021-11-16)
+
+**Note:** Version bump only for package @alfalab/core-components-intl-phone-input
+
+
+
+
+
+## [6.0.1](https://github.com/alfa-laboratory/core-components/compare/@alfalab/core-components-intl-phone-input@6.0.0...@alfalab/core-components-intl-phone-input@6.0.1) (2021-11-08)
+
+**Note:** Version bump only for package @alfalab/core-components-intl-phone-input
+
+
+
+
+
 # [6.0.0](https://github.com/alfa-laboratory/core-components/compare/@alfalab/core-components-intl-phone-input@5.3.3...@alfalab/core-components-intl-phone-input@6.0.0) (2021-10-15)